+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.3] - 12/10/2023
+### Fixed
+- Added support for sendgrid links to the URL Expander. This includes adding wildcard support to the URL Expander.
+- Fixed URL Expanders window to open centred to its parent.
+
 ## [1.5.2] - 21/09/2023
 ### Fixed
 - Fixed exception in filtering for browsers without a logged in user.